added insertrecord api to postgres db
