+++ server/access.log
@@ -119,3 +119,12 @@
 ::ffff:127.0.0.1 - - [14/Aug/2019:19:35:56 +0000] "DELETE /checkout HTTP/1.1" 200 127
 ::1 - - [17/Aug/2019:19:35:40 +0000] "GET / HTTP/1.1" 304 -
 ::1 - - [17/Aug/2019:19:35:40 +0000] "GET /bundle.js HTTP/1.1" 200 4138518
+::1 - - [21/Aug/2019:14:24:45 +0000] "GET / HTTP/1.1" 304 -
+::1 - - [21/Aug/2019:14:24:45 +0000] "GET /bundle.js HTTP/1.1" 200 4138518
+::1 - - [21/Aug/2019:14:24:49 +0000] "GET /checkout HTTP/1.1" 500 1735
+::1 - - [21/Aug/2019:14:24:52 +0000] "POST /checkout HTTP/1.1" 500 1865
+::ffff:127.0.0.1 - - [21/Aug/2019:14:43:54 +0000] "POST /attendees HTTP/1.1" 404 149
+::ffff:127.0.0.1 - - [21/Aug/2019:14:44:56 +0000] "POST /checkout HTTP/1.1" 500 116
+::ffff:127.0.0.1 - - [21/Aug/2019:14:45:42 +0000] "POST /checkout HTTP/1.1" 500 116
+::ffff:127.0.0.1 - - [21/Aug/2019:14:52:10 +0000] "POST /checkout HTTP/1.1" 500 116
+::ffff:127.0.0.1 - - [21/Aug/2019:14:55:44 +0000] "POST /checkout HTTP/1.1" 200 2
